github-318: separate performance tests from the main test suite
